refactor : 공통으로 사용하는 파일들을 밖으로 옮김
